+++ app/src/main/java/com/example/basemvvmexample/data/model/DogImage.kt
@@ -1,2 +1,3 @@
-package com.example.basemvvmexample.data.model 
+package com.example.basemvvmexample.data.model
 
+data class DogImage(val message: String, val status: String)
